remove Deferred handling from BrowserTab
